player moves colect players moves and push them to arbiter
