+++ src/test/java/player/PlayerMovesTest.java
@@ -1,0 +1,7 @@
+package player;
+
+import static org.testng.Assert.*;
+
+public class PlayerMovesTest {
+
+}
